Refactor download script
