+++ scripts/utils/distances.py
@@ -93,3 +93,20 @@ def compute_and_save_station_distances(force_download: bool = False) -> None:
         compression=PARQUET_COMPRESSION,
     )
     print(f"Wrote {distances_df.height} station-pair distances to {STATION_DISTANCES_PATH}")
+
+def enrich_with_distances(rides: pl.LazyFrame, distances: pl.LazyFrame) -> pl.LazyFrame:
+    """"""
+    rides_norm = rides.with_columns([
+        pl.min_horizontal("start_station_id", "end_station_id").alias("_station_min"),
+        pl.max_horizontal("start_station_id", "end_station_id").alias("_station_max"),
+    ])
+    distances_sel = distances.select([
+        pl.col("station_id_a").alias("_station_min"),
+        pl.col("station_id_b").alias("_station_max"),
+        "distance_km",
+    ])
+    return (
+        rides_norm
+        .join(distances_sel, on=["_station_min", "_station_max"], how="left")
+        .drop(["_station_min", "_station_max"])
+    )
