+++ scripts/utils/loaders/hourly_stats.py
@@ -1,0 +1,48 @@
+import polars as pl
+
+def insert_stats_hourly(conn, rides: pl.DataFrame) -> None:
+    """
+    Compute and insert hourly stats into stats_hourly table.
+    Arguments:
+        - conn: psycopg2 connection object to the database
+        - rides: Polars DataFrame containing ride data with columns:
+            - date (datetime)
+            - hour (int)
+            - day_of_week (int)
+            - member_casual (str)
+            - rideable_type (str)
+    """    
+    # Get the count, total duration, and total distance of rides for each date/hour/user_type/bike_type combination
+    agg = (
+        rides
+        .group_by(["date", "hour", "day_of_week", "member_casual", "rideable_type"])
+        .agg([
+            pl.len().alias("total_rides"),
+            pl.col("trip_duration_seconds").sum().alias("total_duration_seconds"),
+            pl.col("distance_km").sum().alias("total_distance_km"),
+        ])
+    )
+    
+    # Convert the aggregated results to a list of tuples for insertion, ensuring proper types and handling nulls
+    rows = [
+        (
+            r["date"], r["hour"], r["day_of_week"], r["member_casual"], r["rideable_type"],
+            int(r["total_rides"]),
+            float(r["total_duration_seconds"] or 0.0),
+            float(r["total_distance_km"] or 0.0),
+        )
+        for r in agg.iter_rows(named=True)
+    ]
+
+    # Insert the aggregated stats into the database
+    with conn.cursor() as cur:
+        cur.executemany(
+            """
+            INSERT INTO stats_hourly
+                (date, hour, day_of_week, user_type, bike_type,
+                 total_rides, total_duration_seconds, total_distance_km)
+            VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
+            """,
+            rows,
+        )
+    print(f"    stats_hourly: {len(rows)} rows")
