+++ scripts/utils/pg_loader.py
@@ -4,20 +4,16 @@ Replaces daily_stats.py — writes three tables:
   stats_hourly, station_activity_hourly, flow_activity_monthly
 plus station_metadata and dataset_coverage.
 """
-from datetime import date
-
 import polars as pl
 
 from src.backend.config import RIDES_DATA_DIR, STATION_DISTANCES_PATH, WEATHER_DATA_DIR
 from src.backend.services.gbfs import fetch_station_data
-from utils.pg_loader_parts.enrichments import _enrich_with_distances
-from utils.pg_loader_parts.inserts import (
-	insert_flow_activity_monthly,
-	insert_station_activity_hourly,
-	insert_stats_hourly,
-	upsert_station_metadata,
-	upsert_weather_hourly,
-)
+from utils.distances import enrich_with_distances
+from utils.loaders.flow_activity_monthly import insert_flow_activity_monthly
+from utils.loaders.hourly_stats import insert_stats_hourly
+from utils.loaders.station_activity_hourly import insert_station_activity_hourly
+from utils.loaders.station_metadata import upsert_station_metadata
+from utils.loaders.weather_hourly import upsert_weather_hourly
 
 def init_db(conn) -> None:
 	"""
@@ -39,7 +35,6 @@ def init_db(conn) -> None:
 	conn.commit()
 	print(f"Database schema initialised from {schema_dir}.")
 
-
 def load_stats_for_month(conn, year: int, month: int) -> None:
 	"""Precompute and insert all stats tables for a single calendar month.."""
 	
@@ -49,7 +44,7 @@ def load_stats_for_month(conn, year: int, month: int) -> None:
 
     # Check if the distance lf is present, join it to rides
 	if STATION_DISTANCES_PATH.exists():
-		rides_lf = _enrich_with_distances(rides_lf, pl.scan_parquet(str(STATION_DISTANCES_PATH)))
+		rides_lf = enrich_with_distances(rides_lf, pl.scan_parquet(str(STATION_DISTANCES_PATH)))
 	else:
 		print(f"Warning: {STATION_DISTANCES_PATH} not found, skipping distance enrichment.")
 		rides_lf = rides_lf.with_columns(pl.lit(None).cast(pl.Float32).alias("distance_km"))
@@ -64,7 +59,6 @@ def load_stats_for_month(conn, year: int, month: int) -> None:
 	conn.commit()
 	print(f"  Done — {year}-{month:02d} committed.")
 
-
 def load_weather_hourly(conn) -> None:
 	"""Load hourly weather parquet data into weather_hourly table."""
 	if not WEATHER_DATA_DIR.exists() or not any(WEATHER_DATA_DIR.rglob("*.parquet")):
@@ -74,7 +68,6 @@ def load_weather_hourly(conn) -> None:
 	weather_df = pl.scan_parquet(str(WEATHER_DATA_DIR / "**/*.parquet"), hive_partitioning=True).collect()
 	upsert_weather_hourly(conn, weather_df)
 	conn.commit()
-
 
 def upsert_station_metadata_from_gbfs(conn) -> None:
 	"""Fetch current station list from GBFS and upsert into station_metadata."""
