--- scripts/utils/pg_loader_parts/enrichments.py
@@ -1,31 +1,0 @@
-import polars as pl
-
-def _enrich_with_distances(rides: pl.LazyFrame, distances: pl.LazyFrame) -> pl.LazyFrame:
-    """Given a LazyFrame of rides and a LazyFrame of pairwise station distances, join to add distance_km to each ride."""
-    rides_norm = rides.with_columns([
-        pl.min_horizontal("start_station_id", "end_station_id").alias("_station_min"),
-        pl.max_horizontal("start_station_id", "end_station_id").alias("_station_max"),
-    ])
-    distances_sel = distances.select([
-        pl.col("station_id_a").alias("_station_min"),
-        pl.col("station_id_b").alias("_station_max"),
-        "distance_km",
-    ])
-    return (
-        rides_norm
-        .join(distances_sel, on=["_station_min", "_station_max"], how="left")
-        .drop(["_station_min", "_station_max"])
-    )
-
-def _enrich_with_weather_code(rides: pl.LazyFrame, weather: pl.LazyFrame) -> pl.LazyFrame:
-    """Given the lazy rides and weather frames, join to add weather_code to each ride based on the hour of the ride."""
-    weather_hourly = weather.select([
-        pl.col("datetime").dt.truncate("1h").alias("_hour_ts"),
-        pl.col("weather_code"),
-    ])
-    return (
-        rides
-        .with_columns(pl.col("started_at").dt.truncate("1h").alias("_hour_ts"))
-        .join(weather_hourly, on="_hour_ts", how="left")
-        .drop("_hour_ts")
-    )
