--- scripts/utils/pg_loader_parts/inserts.py
@@ -1,231 +1,0 @@
-import polars as pl
-
-def insert_stats_hourly(conn, rides: pl.DataFrame) -> None:
-    """Aggregate rides by date, hour, and user/bike type, then insert into stats_hourly."""
-    
-    # Get the count, total duration, and total distance of rides for each date/hour/user_type/bike_type combination
-    agg = (
-        rides
-        .group_by(["date", "hour", "day_of_week", "member_casual", "rideable_type"])
-        .agg([
-            pl.len().alias("total_rides"),
-            pl.col("trip_duration_seconds").sum().alias("total_duration_seconds"),
-            pl.col("distance_km").sum().alias("total_distance_km"),
-        ])
-    )
-    
-    # Convert the aggregated results to a list of tuples for insertion, ensuring proper types and handling nulls
-    rows = [
-        (
-            r["date"], r["hour"], r["day_of_week"], r["member_casual"], r["rideable_type"],
-            int(r["total_rides"]),
-            float(r["total_duration_seconds"] or 0.0),
-            float(r["total_distance_km"] or 0.0),
-        )
-        for r in agg.iter_rows(named=True)
-    ]
-
-    # Insert the aggregated stats into the database
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO stats_hourly
-                (date, hour, day_of_week, user_type, bike_type,
-                 total_rides, total_duration_seconds, total_distance_km)
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
-            ON CONFLICT (date, hour, user_type, bike_type) DO NOTHING
-            """,
-            rows,
-        )
-    print(f"    stats_hourly: {len(rows)} rows")
-
-
-def upsert_weather_hourly(conn, weather_df: pl.DataFrame) -> None:
-    """Upsert hourly weather into weather_hourly table."""
-    if weather_df.is_empty():
-        print("    weather_hourly: 0 rows upserted")
-        return
-
-    weather_hourly = (
-        weather_df
-        .with_columns([
-            pl.col("datetime").dt.date().alias("date"),
-            pl.col("datetime").dt.hour().cast(pl.Int16).alias("hour"),
-        ])
-        .select([
-            "date",
-            "hour",
-            pl.col("weather_code").cast(pl.Int16),
-            pl.col("temperature_2m").cast(pl.Float64),
-            pl.col("precipitation").cast(pl.Float64),
-            pl.col("wind_speed_10m").cast(pl.Float64),
-        ])
-        .unique(subset=["date", "hour"], keep="last")
-    )
-
-    rows = [
-        (
-            r["date"],
-            int(r["hour"]),
-            int(r["weather_code"]) if r["weather_code"] is not None else None,
-            float(r["temperature_2m"]) if r["temperature_2m"] is not None else None,
-            float(r["precipitation"]) if r["precipitation"] is not None else None,
-            float(r["wind_speed_10m"]) if r["wind_speed_10m"] is not None else None,
-        )
-        for r in weather_hourly.iter_rows(named=True)
-    ]
-
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO weather_hourly
-                (date, hour, weather_code, temperature_2m, precipitation, wind_speed_10m)
-            VALUES (%s, %s, %s, %s, %s, %s)
-            ON CONFLICT (date, hour) DO UPDATE SET
-                weather_code   = EXCLUDED.weather_code,
-                temperature_2m = EXCLUDED.temperature_2m,
-                precipitation  = EXCLUDED.precipitation,
-                wind_speed_10m = EXCLUDED.wind_speed_10m
-            """,
-            rows,
-        )
-    print(f"    weather_hourly: {len(rows)} rows upserted")
-
-def insert_station_activity_hourly(conn, rides: pl.DataFrame) -> None:
-    """Aggregate rides by month, day-of-week, hour, and station, then insert into station_activity_hourly."""
-    rides = rides.with_columns([
-        pl.col("date").dt.year().cast(pl.Int16).alias("year"),
-        pl.col("date").dt.month().cast(pl.Int16).alias("month"),
-    ])
-    key_cols = ["year", "month", "day_of_week", "hour", "member_casual", "rideable_type"]
-    outgoing = (
-        rides
-        .group_by([*key_cols, "start_station_id"])
-        .agg(pl.len().alias("outgoing_rides"))
-        .rename({"start_station_id": "station_id"})
-    )
-    incoming = (
-        rides
-        .group_by([*key_cols, "end_station_id"])
-        .agg(pl.len().alias("incoming_rides"))
-        .rename({"end_station_id": "station_id"})
-    )
-    activity = (
-        outgoing
-        .join(incoming, on=[*key_cols, "station_id"], how="full", coalesce=True)
-        .with_columns([
-            pl.col("outgoing_rides").fill_null(0),
-            pl.col("incoming_rides").fill_null(0),
-        ])
-    )
-    rows = [
-        (
-            int(r["year"]), int(r["month"]), int(r["day_of_week"]), int(r["hour"]),
-            r["station_id"], r["member_casual"], r["rideable_type"],
-            int(r["outgoing_rides"]), int(r["incoming_rides"]),
-        )
-        for r in activity.iter_rows(named=True)
-    ]
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO station_activity_hourly
-                (year, month, day_of_week, hour, station_id, user_type, bike_type, outgoing_rides, incoming_rides)
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)
-            ON CONFLICT (year, month, day_of_week, hour, station_id, user_type, bike_type) DO NOTHING
-            """,
-            rows,
-        )
-    print(f"    station_activity_hourly: {len(rows)} rows")
-
-def insert_flow_activity_monthly(conn, rides: pl.DataFrame) -> None:
-    """Aggregate rides by month, station pair, user/bike type, and insert flow counts into flow_activity_monthly."""
-    flow = (
-        rides
-        .with_columns([
-            pl.col("date").dt.year().cast(pl.Int16).alias("year"),
-            pl.col("date").dt.month().cast(pl.Int16).alias("month"),
-            pl.min_horizontal("start_station_id", "end_station_id").alias("station_a_id"),
-            pl.max_horizontal("start_station_id", "end_station_id").alias("station_b_id"),
-            (pl.col("start_station_id") <= pl.col("end_station_id")).alias("_is_a_to_b"),
-        ])
-        .group_by(["year", "month", "station_a_id", "station_b_id", "member_casual", "rideable_type", "_is_a_to_b"])
-        .agg(pl.len().alias("count"))
-    )
-    key_cols = ["year", "month", "station_a_id", "station_b_id", "member_casual", "rideable_type"]
-    a_to_b = flow.filter(pl.col("_is_a_to_b")).drop("_is_a_to_b").rename({"count": "a_to_b_count"})
-    b_to_a = flow.filter(~pl.col("_is_a_to_b")).drop("_is_a_to_b").rename({"count": "b_to_a_count"})
-    merged = (
-        a_to_b
-        .join(b_to_a, on=key_cols, how="full", coalesce=True)
-        .with_columns([
-            pl.col("a_to_b_count").fill_null(0),
-            pl.col("b_to_a_count").fill_null(0),
-        ])
-    )
-    rows = [
-        (
-            int(r["year"]), int(r["month"]), r["station_a_id"], r["station_b_id"],
-            r["member_casual"], r["rideable_type"],
-            int(r["a_to_b_count"]), int(r["b_to_a_count"]),
-        )
-        for r in merged.iter_rows(named=True)
-    ]
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO flow_activity_monthly
-                (year, month, station_a_id, station_b_id, user_type, bike_type, a_to_b_count, b_to_a_count)
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
-            ON CONFLICT (year, month, station_a_id, station_b_id, user_type, bike_type) DO NOTHING
-            """,
-            rows,
-        )
-    print(f"    flow_activity_monthly: {len(rows)} rows")
-
-def upsert_station_metadata(conn, station_info: list[dict]) -> None:
-    """Upsert station metadata from GBFS feed into station_metadata table."""
-    rows = [
-        (s["short_name"], s["name"], s["lat"], s["lon"])
-        for s in station_info
-    ]
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO station_metadata (station_id, station_name, lat, lon)
-            VALUES (%s, %s, %s, %s)
-            ON CONFLICT (station_id) DO NOTHING
-            """,
-            rows,
-        )
-    print(f"    station_metadata: {len(rows)} stations upserted")
-
-def upsert_bike_routes(conn, df: pl.DataFrame) -> None:
-    """Upsert all current bike route segments into the bike_routes table."""
-    df = df.with_columns(
-        pl.col("instdate").str.strptime(pl.Date, "%m/%d/%Y", strict=False)
-    )
-    rows = [
-        (int(r["segmentid"]), int(r["bikeid"]), r["the_geom"], r["street"], r["fromstreet"],
-         r["tostreet"], r["facilitycl"], r["instdate"], r["boro"])
-        for r in df.iter_rows(named=True)
-    ]
-    with conn.cursor() as cur:
-        cur.executemany(
-            """
-            INSERT INTO bike_routes
-                (segmentid, bikeid, the_geom, street, fromstreet, tostreet, facilitycl, instdate, boro)
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)
-            ON CONFLICT (segmentid) DO UPDATE SET
-                bikeid     = EXCLUDED.bikeid,
-                the_geom   = EXCLUDED.the_geom,
-                street     = EXCLUDED.street,
-                fromstreet = EXCLUDED.fromstreet,
-                tostreet   = EXCLUDED.tostreet,
-                facilitycl = EXCLUDED.facilitycl,
-                instdate   = EXCLUDED.instdate,
-                boro       = EXCLUDED.boro
-            """,
-            rows,
-        )
-    print(f"    bike_routes: {len(rows)} rows upserted")
