+++ src/backend/routes/stats.py
@@ -38,6 +38,24 @@ def get_stats(
         end_station_id=end_station_id,
     )
 
+@router.get("/stats_by_weather", response_model=list[GroupedStats])
+def get_stats_by_weather(
+    group_by: StatsGroupBy = Query(default=StatsGroupBy.NONE),
+    user_type: MemberCasual | None = Query(default=None),
+    bike_type: RideableType | None = Query(default=None),
+    start_date: date = Query(...),
+    end_date: date = Query(...),
+):
+    """Get historical rides stats joined with weather conditions, optionally grouped by day_of_week, hour, or both."""
+    return get_stats_data(
+        group_by=group_by,
+        user_type=user_type,
+        bike_type=bike_type,
+        start_date=start_date,
+        end_date=end_date,
+        join_weather=True,
+    )
+
 @router.get("/station_usage_counts", response_model=list[StationRideCounts])
 def get_station_ride_counts(
     group_by: RideCountGroupBy = Query(default=RideCountGroupBy.NONE),
